SSE-3118:Reintroduce GA4 Code into Main Branch
