+++ .env.example
@@ -33,3 +33,10 @@ PORT=3000
 USE_STUB_SHEETS=true
 USE_STUB_ZENDESK=true
 USE_STUB_SERVICENOW=true
+
+# Google Analytics
+GOOGLE_ANALYTICS_4_GTM_CONTAINER_ID=somevalue
+UNIVERSAL_ANALYTICS_GTM_CONTAINER_ID=somevalue
+GA4_DISABLED=somevalue
+UA_DISABLED=somevalue
+COOKIE_DOMAIN=thedomainthatapprunson
