+++ package.json
@@ -29,6 +29,7 @@
   "license": "ISC",
   "dependencies": {
     "@googleapis/sheets": "^0.3.0",
+    "@govuk-one-login/one-login-analytics": "^0.0.9",
     "axios": "^1.6.2",
     "email-validator": "^2.0.4",
     "express": "^4.19.2",
