+++ src/app.ts
@@ -10,6 +10,7 @@ import redirects from "./routes/redirects";
 import register from "./routes/register";
 import site from "./routes/site";
 import support from "./routes/support";
+import path from "path";
 
 const app = express();
 app.set("validation", Validation.getInstance());
@@ -25,6 +26,7 @@ app.use(Helmet());
 
 app.use("/assets", express.static(distribution.assets));
 app.use("/assets/images", express.static(distribution.images));
+app.use("/ga4-assets", express.static(path.join(__dirname, "../node_modules/@govuk-one-login/one-login-analytics/lib")));
 
 app.use(express.urlencoded()); // to support URL-encoded bodies
 app.use(
@@ -43,9 +45,13 @@ app.use("/", redirects);
 app.use("/", support);
 app.use("/", mailingList);
 
-app.locals.googleTagId = process.env.GOOGLE_TAG_ID;
 app.locals.showTestBanner = process.env.SHOW_TEST_BANNER === "true";
 app.locals.adminToolUrl = process.env.ADMIN_TOOL_URL || "https://admin.sign-in.service.gov.uk";
+app.locals.ga4ContainerId = process.env.GOOGLE_ANALYTICS_4_GTM_CONTAINER_ID;
+app.locals.uaContainerId = process.env.UNIVERSAL_ANALYTICS_GTM_CONTAINER_ID;
+app.locals.ga4Disabled = process.env.GA4_DISABLED;
+app.locals.uaDisabled = process.env.UA_DISABLED;
+app.locals.cookieDomain = process.env.COOKIE_DOMAIN || "sign-in.service.gov.uk";
 
 app.use((req: Request, res: Response) => {
     res.status(404).render("404.njk");
