+++ src/views/about.njk
@@ -44,4 +44,16 @@
   <h3 class="govuk-heading-s">Support</h3>
   <p class="govuk-body">Support for your service is available during office hours via a <a class="govuk-link" href="/contact-us ">support form</a> and <a class="govuk-link" href="https://ukgovernmentdigital.slack.com/archives/C02AQUJ6WTC">Slack channel</a>. We aim to reply within 2 working days.</p>
   <p class="govuk-body">If you go live with GOV.UK One Login, our technical team is available 24/7 for urgent incidents.</p>
+
+{% endblock %}
+{% block scripts %}
+  {{ga4OnPageLoad({
+    statusCode:"200",
+    englishPageTitle: pageTitle,
+    taxonomyLevel1: taxonomyLevel1,
+    taxonomyLevel2: taxonomyLevel2About,
+    contentId:"5ef8618c-007b-4aa0-9093-873fca11a266",
+    loggedInStatus:false,
+    dynamic:false
+  })}}
 {% endblock %}
