+++ src/views/base.njk
@@ -1,5 +1,6 @@
 {% extends "govuk/template.njk" %}
 {% from "govuk/components/phase-banner/macro.njk" import govukPhaseBanner %}
+{% from "macros/ga4-opl.njk" import ga4OnPageLoad %}
 
 {% set assetPath = "/assets" %}
 {% set documentationUrl = "/documentation" %}
@@ -10,6 +11,14 @@
 {% set phaseTagText  = "test" if showTestBanner else "beta" %}
 {% set phaseBannerClasses  = "test-banner" if showTestBanner %}
 {% set phaseBannerHtml = "This is a test environment." %}
+
+{% set taxonomyLevel1 = "adoption" %}
+{% set taxonomyLevel2About = "about" %}
+{% set taxonomyLevel2Documents = "documents" %}
+{% set taxonomyLevel2Register = "register" %}
+{% set taxonomyLevel2Support = "support" %}
+{% set taxonomyLevel2Policies = "policies" %}
+
 {% if not showTestBanner %}
   {% set phaseBannerHtml %}
     This is a new service – your
@@ -164,13 +173,22 @@
 {% endblock %}
 
 {% block bodyEnd %}
-  <input type="hidden" value="{{ googleTagId }}" id="ga-tracking">
-  {# Run JavaScript at end of the <body>, to avoid blocking the initial render. #}
+  <input type="hidden" value="{{cookieDomain}}" id="ga-cookie-domain">
   <script src="/assets/javascripts/govuk-all.js"></script>
-  <script src="/assets/javascripts/cookies.js"></script>
+  <script src="/ga4-assets/analytics.js"></script>
+  <script src="/assets/javascripts/ua.js"></script>
+  {% block analyticscripts %}{% endblock %}
   <script>window.GOVUKFrontend.initAll()</script>
-  <script>window.GOVUKSignIn.CookieBanner.init()</script>
-
+  <script>
+    window.DI.appInit({
+        ga4ContainerId: "{{ga4ContainerId}}",
+        uaContainerId: "{{uaContainerId}}"
+      }, {
+        disableGa4Tracking: {{ga4Disabled}},
+        disableUaTracking: {{uaDisabled}},
+        cookieDomain: "{{cookieDomain}}"
+      });
+  </script>
   {% block scripts %}{% endblock %}
   {% block formScripts %}{% endblock %}
 {% endblock %}
