+++ src/views/common/cookie-banner.njk
@@ -14,7 +14,7 @@
 
 {{ govukCookieBanner({
   ariaLabel: "Cookies on GOV.UK One Login",
-  classes: "govuk-!-display-none",
+  name: "gov-uk-cookie-banner",
   messages: [
     {
       headingText: "Cookies on GOV.UK One Login",
@@ -26,16 +26,14 @@
         {
           text: "Accept analytics cookies",
           type: "button",
-          attributes: {
-            id: "cookiesAccept"
-          }
+          name: "cookiesAccept",
+          value: "accept"
         },
         {
           text: "Reject analytics cookies",
           type: "button",
-          attributes: {
-            id: "cookiesReject"
-          }
+          name: "cookiesReject",
+          value: "reject"
         },
         {
           text: "View cookies",
@@ -46,30 +44,34 @@
     {
       html: acceptMessageHtml,
       role: "alert",
-      classes: "govuk-!-display-none",
       attributes: {
         id: "cookies-accepted"
       },
       actions: [
         {
           text: "Hide this message",
-          classes: "cookie-hide-button"
+          classes: "cookie-hide-button",
+          href: "#",
+          type: "button"
         }
-      ]
+      ],
+      hidden: true
     },
     {
       html: rejectMessageHtml,
       role: "alert",
-      classes: "govuk-!-display-none",
         attributes: {
         id: "cookies-rejected"
       },
       actions: [
         {
           text: "Hide this message",
-          classes: "cookie-hide-button"
+          classes: "cookie-hide-button",
+          href: "#",
+          type: "button"
         }
-      ]
+      ],
+      hidden: true
     }
-  ]
+  ], hidden: false
 }) }}
