+++ src/views/contact-us-confirm.njk
@@ -18,3 +18,14 @@
     Return to the <a class="govuk-link" href="/">GOV.UK One Login homepage</a>.
   </p>
 {% endblock %}
+{% block scripts %}
+  {{ga4OnPageLoad({
+    statusCode:"200",
+    englishPageTitle: pageTitle,
+    taxonomyLevel1: taxonomyLevel1,
+    taxonomyLevel2: taxonomyLevel2Support,
+    contentId:"920f681c-dbc1-4481-bdae-2d9c1f72b0d4",
+    loggedInStatus:false,
+    dynamic:false
+  })}}
+{% endblock %}
