+++ src/views/contact-us-details.njk
@@ -22,3 +22,14 @@
     Or if you still have questions, you can contact us using our <a class="govuk-link" href="/contact-us ">support form</a> or via our <a class="govuk-link" href="https://ukgovernmentdigital.slack.com/archives/C02AQUJ6WTC">Slack channel</a>.
   </p>
 {% endblock %}
+{% block scripts %}
+  {{ga4OnPageLoad({
+    statusCode:"200",
+    englishPageTitle: pageTitle,
+    taxonomyLevel1: taxonomyLevel1,
+    taxonomyLevel2: taxonomyLevel2Support,
+    contentId:"f27709e0-5c6d-4f7b-8dbe-478849fbd493",
+    loggedInStatus:false,
+    dynamic:false
+  })}}
+{% endblock %}
