+++ src/views/documentation-change-credentials.njk
@@ -96,4 +96,13 @@
 
 {% block scripts %}
   <script type="module" src="/assets/javascripts/back-to-top.js"></script>
+  {{ga4OnPageLoad({
+    statusCode:"200",
+    englishPageTitle: pageTitle,
+    taxonomyLevel1: taxonomyLevel1,
+    taxonomyLevel2: taxonomyLevel2Documents,
+    contentId:"0b9f68f4-30fc-40d8-941d-6f61bce02dff",
+    loggedInStatus:false,
+    dynamic:false
+  })}}
 {% endblock %}
