+++ src/views/macros/ga4-opl.njk
@@ -1,0 +1,3 @@
+{% macro ga4OnPageLoad(params) %}
+    {%- include "./ga4-opl-template.njk" -%}
+{% endmacro %}
