+++ src/views/support.njk
@@ -34,3 +34,14 @@
   <p class="govuk-body">We can only respond to your support queries during office support hours. Our support hours are 8:00am to 8:00pm, Monday to Friday.</p>
   <p class="govuk-body">When you get in touch, we’ll try to reply within 2 working days.</p>
 {% endblock %}
+{% block scripts %}
+  {{ga4OnPageLoad({
+    statusCode:"200",
+    englishPageTitle: pageTitle,
+    taxonomyLevel1: taxonomyLevel1,
+    taxonomyLevel2: taxonomyLevel2Support,
+    contentId:"94096541-db52-444c-9b02-098e1622e581",
+    loggedInStatus:false,
+    dynamic:false
+  })}}
+{% endblock %}
